requirements.txt
Flask
python-telegram-bot==20.0
gunicorn
aiohttp
json
asyncio